좋아요 등록, 조회 구현

+++ controllers/likes.controller.js
@@ -1,27 +1,45 @@
-const { urlencoded } = require('express');
 const LikesService = require('../services/likes.service');
 
 class LikesController {
   likesService = new LikesService();
   putLikes = async (req, res, next) => {
-    try {
-      const { postId } = req.params;
-      const { userId } = res.locals.user;
+    const { postId } = req.params;
+    const { userId } = res.locals.user;
 
-      const existsLike = await this.likesService.findLike(postId, userId);
+    const existsLike = await this.likesService.findLike(postId, userId);
+    const existsPost = await this.likesService.findOnePost(postId);
+
+    if (!existsPost) {
+      throw new Error('404, 게시글이 존재하지않습니다.');
+    }
+
+    try {
       if (existsLike) {
         console.log(existsLike);
         await this.likesService.deleteLikes(postId, userId);
+        await this.likesService.dropLike(postId);
         return res.status(201).json({
           message: '좋아요를 취소했습니다.',
         });
       }
 
       const createLike = await this.likesService.putLikes(postId, userId);
+      await this.likesService.addLike(postId);
       res.status(201).json({ message: '좋아요를 추가했습니다.', createLike });
     } catch (err) {
       console.error(err);
       res.status(400).json({ message: '게시글 좋아요에 실패하였습니다.' });
+    }
+  };
+
+  getLikes = async (req, res, next) => {
+    const { userId } = res.locals.user;
+    try {
+      const posts = await this.likesService.findPostLikes(userId);
+      return res.status(201).json({ posts });
+    } catch (error) {
+      console.error(error);
+      res.status(400).json({ message: '좋아요 게시글 조회에 실패하였습니다.' });
     }
   };
 }
@@ -29,12 +47,3 @@ class LikesController {
 //---->
 
 module.exports = LikesController;
-
-//이미 좋아요가 추가되어있는지 확인
-//   const existsLike = await this.likesService.putLikes(postId, userId);
-//   if (existsLike) {
-//     await this.likesService.deleteLikes(postId, userId);
-//     return res.status(200).json({ message: '좋아요를 취소하였습니다.' });
-//   }
-
-//좋아요가 없으면 좋아요를 등록
